Changed from O(N^2) to O(N*LOG(2,N))
In most Javascript interpreters, in concatenation, in the creation of the result string, each character of each string must be iterated through. In the old version, there was a loop of each required iteration. As a result, with N concatenations and N characters, the performance is O(N^2). In this version, the number of concatenations is logarithmic to the number of strings to be combined, with a base of somewhere between 2 and 3. As a result, this is O(N*LOG(K,N)). Though it requires a few kilobytes more memory, this is a major performance improvement at large scales.