index.js changes
I think I fixed the issue with padding with multi-character strings. It tested correctly on a linux and windows machine. Also, the performance is now O(N*LOG(K,N)), instead of O(N^2). When Javascript performs string concatenations, it has to iterate through each character. Since the old version iterates through them in a loop, the characters are iterated N times within N times, being N^2. This version reduces the number of operations to a hybrid log of 2 and 3.